Break database into three classes to add subsonic support
